functions/update_hints/requirements.txt
beautifulsoup4==4.11.1
google-cloud-storage
requests==2.28.1
